Bug 12109: Fix tab on the auth subfield configuration page
There is a condition in the template to not display the "New" tab if the
user is editing a control field.
But it results in a broken template.
Looking at the biblio frameworks, the New tab exists even for control
fields.

This patch only fixes the html structure, but a complete fix should be
provide, for auth and biblio, if we want not to let the ability to
create subfields for control fields.

Test plan:
Edit subfields for a control field
(admin/marc_subfields_structure.pl?op=add_form&tagfield=001&frameworkcode= for instance)
and confirm that the view is now fixed.

Signed-off-by: Mark Tompsett <mtompset@hotmail.com>

NOTE: Checked LDR, 001, and 008.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>